Extraction/chunking_based_on_llms/src/requirements.txt
pymupdf==1.23.15
gspread==6.0.0
langchain==0.1.4
langchain_community==0.0.16
langchain_core==0.1.16
langchain_nvidia_ai_endpoints==0.0.1
pandas==2.2.0
Pillow==10.2.0
pydantic==2.5.3
pymilvus==2.3.5
python_pptx==0.6.23
qdrant_client==1.7.1
Requests==2.31.0
torch==2.2.0
transformers==4.35.2
nltk
pypdf
fuzzywuzzy
openai
anthropic

fitz
frontend